Backend_with_input/requirements.txt
numpy
pandas
scipy
openpyxl
refinitiv-data
typing